ExitAirSensor: Get power readings from PSU
Now that PSU Sensors work, remove todo. Also fix
a bug where a int32 should have been a double as the
value from d-bus is a double.

Tested: Ipmitool sensor list

+++ include/ExitAirTempSensor.hpp
@@ -10,9 +10,9 @@ struct ExitAirTempSensor;
 struct CFMSensor : public Sensor
 {
     std::vector<std::string> tachs;
-    int32_t c1;
-    int32_t c2;
-    int32_t maxCFM;
+    double c1;
+    double c2;
+    double maxCFM;
     double tachMinPercent;
     double tachMaxPercent;
